+++ backend/src/main/java/com/example/backend/model/BettingRound.java
@@ -8,7 +8,7 @@ import java.util.Map;
 
 @Data
 public class BettingRound {
-    private Map<String, Double> bets = new HashMap<>();
+    private Map<String, Long> bets = new HashMap<>();
     private RoundType roundType;
     private boolean roundComplete;
