+++ backend/src/main/java/com/example/backend/service/GameValidatorService.java
@@ -92,16 +92,16 @@ public class GameValidatorService {
     }
 
     public void validateCanCheck(Game game, Player player) {
-        double betPlaced = game.getCurrentBettingRound().getBets().getOrDefault(player.getId(), 0.0);
+        long betPlaced = game.getCurrentBettingRound().getBets().getOrDefault(player.getId(), 0L);
         if (betPlaced < game.getCurrentBet()) {
             logger.error("Cannot check while facing a bet");
             throw new IllegalArgumentException("Cannot check, you must call or fold");
         }
     }
 
-    public void validatePlayerBetAmount(Game game, Player player, double amount) {
-        double betPlacedAmount = game.getCurrentBettingRound().getBets().getOrDefault(player.getId(), 0.0);
-        double betPlacedTotal = betPlacedAmount + amount;
+    public void validatePlayerBetAmount(Game game, Player player, long amount) {
+        long betPlacedAmount = game.getCurrentBettingRound().getBets().getOrDefault(player.getId(), 0L);
+        long betPlacedTotal = betPlacedAmount + amount;
         if (amount < 0) {
             logger.error("Bet amount cannot be negative");
             throw new IllegalArgumentException("Bet amount cannot be negative");
